# openapi.yaml
openapi: 3.0.3
info:
  title: Cache Presentation API
  description: REST API для демонстрации стратегий кэширования Cache-Aside, Write-Through, Write-Behind и Read-Through (Spring + Redis + Caffeine + PostgreSQL)
  version: 0.0.2
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /coupons:
    post:
      tags: [Coupons]
      summary: Создать купон
      description: |
        **Negative Caching CREATE**
        - Сохраняет купон в PostgreSQL
        - `@CacheEvict` сбрасывает закешированный `null` для данного `code`
        - Следующий `GET /{code}` найдёт реальный объект в БД и закеширует его
      operationId: createCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponRequest'
      responses:
        '200':
          description: Созданный купон
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
  /coupons/{code}:
    get:
      tags: [Coupons]
      summary: Получить купон по коду
      description: |
        **Negative Caching READ**
        - HIT (объект) — купон найден в Redis, БД не вызывается → `200`
        - HIT (null)   — в Redis закешировано отсутствие, БД не вызывается → `404`
                         **Именно это предотвращает Cache Penetration.**
        - MISS         — идёт в БД. Результат (объект или `null`) кешируется на 30 сек.

        Проверить в логах: `[DB] Loading coupon 'X'...` появится только при первом
        запросе к несуществующему коду, пока не истечёт TTL.
      operationId: getCouponByCode
      parameters:
        - $ref: '#/components/parameters/CouponCode'
      responses:
        '200':
          description: Купон найден (из Redis или БД)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
        '404':
          description: Купон не найден (null закеширован в Redis — БД не вызывалась)

    delete:
      tags: [Coupons]
      summary: Удалить купон
      description: |
        **Negative Caching DELETE**
        - Удаляет из PostgreSQL
        - `@CacheEvict` сбрасывает кеш
        - Следующий `GET /{code}` снова пойдёт в БД и закеширует `null`
      operationId: deleteCoupon
      parameters:
        - $ref: '#/components/parameters/CouponCode'
      responses:
        '204':
          description: Купон удалён
  /jitter/products/{id}:
    get:
      tags: [Jitter Cache]
      summary: Получить продукт (с TTL + Jitter)
      description: |
        **TTL + Jitter READ** — защита от Cache Stampede

        При cache miss загружает из PostgreSQL и кладёт в Redis с TTL = `60 ± 10 сек`
        (случайный диапазон 50..70 сек).

        **Зачем jitter?**
        Без него все записи, закешированные одновременно, протухают одновременно
        → тысячи запросов разом уходят в БД (thundering herd).
        Случайный разброс TTL размазывает истечение во времени.

        В логах при каждом cache miss виден реальный TTL:
        ```
        [CACHE SET] product 3 → TTL 57 sec (base=60, jitter=-3)
        ```
      operationId: getProductByIdWithJitter
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Продукт найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '500':
          description: Продукт не найден
  /jitter/products/{id}/cache:
    delete:
      tags: [Jitter Cache]
      summary: Сбросить кеш продукта (для демонстрации)
      description: |
        Удаляет ключ `jitter:product:{id}` из Redis.
        После этого следующий `GET /jitter/products/{id}` даст cache miss
        и закеширует продукт с новым случайным TTL.
      operationId: evictProductJitterCache
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Кеш сброшен
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 1
...
    CouponCode:
      name: code
      in: path
      required: true
      schema:
        type: string
        example: SAVE10
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: MacBook Pro
        price:
          type: number
          format: decimal
          example: 199999.99
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        code:
          type: string
          example: SAVE10
        discount:
          type: number
          format: decimal
          example: 10.00
    CouponRequest:
      type: object
      required: [code, discount]
      properties:
        code:
          type: string
          example: GHOST
        discount:
          type: number
          format: decimal
          example: 5.00